Removing the server.go file.
